Move ThrowsIfQueryAttributeWithRawQueryStringAttribute
