+++ src/RestEase.UnitTests/ImplementationFactoryTests/MultipleParameterAttributesTests.cs
@@ -1,4 +1,5 @@
 ﻿using System.Threading.Tasks;
+using RestEase.Implementation;
 using Xunit;
 using Xunit.Abstractions;
 
@@ -10,6 +11,12 @@ namespace RestEase.UnitTests.ImplementationFactoryTests
         {
             [Get("/{bar}")]
             Task FooAsync([Query, Header("header1"), Body, Path, HttpRequestMessageProperty("prop1")] string bar);
+        }
+
+        public interface IHasMethodParameterWithConflictAttributes
+        {
+            [Get]
+            Task FooAsync([Query, RawQueryString] string foo);
         }
 
         public MultipleParameterAttributesTests(ITestOutputHelper output) : base(output) { }
@@ -24,6 +31,17 @@ namespace RestEase.UnitTests.ImplementationFactoryTests
             Assert.NotNull(requestInfo.BodyParameterInfo);
             Assert.Single(requestInfo.PathParams);
             Assert.Single(requestInfo.HttpRequestMessageProperties);
+        }
+
+        [Fact]
+        public void ThrowsIfQueryAttributeWithRawQueryStringAttribute()
+        {
+            VerifyDiagnostics<IHasMethodParameterWithConflictAttributes>(
+                // (4,27): Error REST040: Method 'FooAsync': [Query] parameter must not specified along with [RawQueryString]
+                // [Query, RawQueryString] string foo
+                Diagnostic(DiagnosticCode.QueryConflictWithRawQueryString, @"[Query, RawQueryString] string foo")
+                    .WithLocation(4, 27)
+            );
         }
     }
 }
